+++ DEMO_Out_of_Office_App/DTOS/EmployeeDTO.cs
@@ -1,5 +1,5 @@
 ﻿namespace DEMOOutOfOfficeApp.DTOS
 {
-    public record EmployeeDTO(int ID,string FullName, string SubdivisionName, string PositionName,string StatusName,string PeoplePartnerName, decimal OutOfOfficeBalance, byte[]? Photo,int rolePositionId = 0);
+    public record EmployeeDTO(int ID,string FullName, string SubdivisionName, string PositionName,string StatusName,string PeoplePartnerName, int OutOfOfficeBalance, byte[]? Photo,int rolePositionId = 0);
  
 }
